Fixes to RefinementVariable handling. Added BallSizeTestConstraint handling.
This changes creates a new RefinementVariable when DFF Stores are merged and have conflicting annotations for a Variable. (Either VariableAnnotation and RefinementVariable, or two RefinementVariables). The merge variable is created in the InferenceValue.leastUpperBound method.

Previously, inference annotations were not considered valid values in the DFF system, so they would not be kept between store merges. Now they should be.  This lead to a few issues:
VisitTernary was doing a leastUpperBound between the lhs and rhs result values. We don't want this to happen for RefinementVariables (this is already handled as a CombVariable later).
The default strengthenAnnotationOfEqualTo may replace one RefinementVariable with another (because they subtype relationship between them is not correctly defined).
DFF would chose to use VariableAnnotation over a RefinementVariable because it thought the VariableAnnotation was more specific (because they subtype relationship between them is not correctly defined).

I fixed these issues as the came up. The subtype relationship between inference annotations during flow is still only partly defined (maybe we should clean that up).

RefinementVariables no longer have subtype constraints with the RHS assignment, instead have an equality constraint to the RHS value that was being assigned. The RHS value now has a subtype constraint to the RefinementVariable's Declaration Variable.

The equality constraint removed some of the need for tweaking the order of refinement variables in the game solver, however, some ordering is was left in to merge variables so that it more accurately reflected the underlying code (this flows into that, etc).

NninfTransfer was created to generate BallSizeTestConstraints on equality / inequality with null.

Using a new InferenceValue and new InferenceTransfer has lead to allowing command line arguments for setting Transfers and Analysis.

https://rbcommons.com/s/PLSE/r/58/